# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"paths":{"/v1/user/{id}":{"get":{"tags":["user-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}}}},"put":{"tags":["user-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["user-controller"],"operationId":"deleteById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/v1/user":{"get":{"tags":["user-controller"],"summary":"Get list of users","operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["user-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/v1/user/{email}/{password}":{"get":{"tags":["user-controller"],"operationId":"getByNameAndDescription","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string"}},{"name":"password","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"Person":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"lastname":{"type":"string"},"document":{"type":"string"}}},"UserModel":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"person":{"$ref":"#/components/schemas/Person"},"userId":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date-time"}}}}}}