# event/src/main/java/com/bootcampfinal/java/event/web/model/UserModel.java
package com.bootcampfinal.java.event.web.model;

import com.bootcampfinal.java.event.domain.Person;
import com.fasterxml.jackson.annotation.JsonFormat;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.sql.Timestamp;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UserModel {


    @JsonProperty("userId")
    private Long id;

    private String email;

    private String password;

    @JsonProperty("date")
    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm")
    private Timestamp dateEvent;

}